# packages/protocol/openapi/openapi.json
{
  "openapi": "3.1.1",
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "info": {
    "title": "Formation Chat Core API",
    "version": "0.1.0",
    "description": "Public chat and separately authenticated read-only administration APIs. Writes are ordinary HTTP requests; public conversation events stream over SSE."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "sessionToken": [] }],
  "paths": {
    "/v1/sessions": {
      "post": {
        "operationId": "bootstrapAnonymousSession",
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "../schemas/identity/anonymous-bootstrap-request.schema.json" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created or resumed",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/identity/session-bootstrap-response.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/identity/exchange": {
      "post": {
        "operationId": "exchangeHostIdentity",
        "security": [{ "serviceCredential": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "../schemas/identity/identity-exchange-request.schema.json" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scoped session",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/identity/session-bootstrap-response.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations": {
      "get": {
        "operationId": "listConversations",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated conversations",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/conversation-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "createConversation",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "../schemas/chat/create-conversation-request.schema.json" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/conversation.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "getConversation",
        "responses": {
          "200": {
            "description": "Canonical conversation snapshot",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/conversation.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}/messages": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "listMessages",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated messages",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/message-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "submitMessage",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "../schemas/chat/submit-message-request.schema.json" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Message accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/message.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}/events": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "streamConversationEvents",
        "parameters": [{ "$ref": "#/components/parameters/LastEventId" }],
        "responses": {
          "200": {
            "description": "SSE stream. Each frame uses id=<eventId>, event=<type>, and data=<JSON ConversationEvent>. Only events visible to the session scope are emitted.",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "x-event-payload-schema": {
                  "$ref": "../schemas/chat/public-conversation-event.schema.json"
                }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}/inputs/{requestId}": {
      "parameters": [
        { "$ref": "#/components/parameters/ConversationId" },
        { "$ref": "#/components/parameters/RequestId" }
      ],
      "post": {
        "operationId": "submitStructuredInput",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/chat/submit-structured-input-request.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Input accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/structured-input-request.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}/cancel": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "post": {
        "operationId": "cancelConversationRun",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "responses": {
          "202": {
            "description": "Best-effort cancellation outcome",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/cancel-run-response.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/conversations/{conversationId}/retry": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "post": {
        "operationId": "retryConversationRun",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "responses": {
          "202": { "description": "Retry accepted" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/conversations": {
      "get": {
        "operationId": "adminListConversations",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/AgentRef" },
          { "$ref": "#/components/parameters/CreatedAfter" },
          { "$ref": "#/components/parameters/CreatedBefore" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["active", "completed", "cancelled"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin-scoped conversations",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/conversation-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/conversations/{conversationId}": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "adminGetConversation",
        "security": [{ "adminToken": [] }],
        "responses": {
          "200": {
            "description": "Admin-scoped conversation",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/chat/conversation.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/conversations/{conversationId}/messages": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "adminListMessages",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Admin-scoped canonical messages",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/message-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/conversations/{conversationId}/events": {
      "parameters": [{ "$ref": "#/components/parameters/ConversationId" }],
      "get": {
        "operationId": "adminListEvents",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Event timeline filtered by admin visibility scope",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/event-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/runs": {
      "get": {
        "operationId": "adminListRuns",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/AgentRef" },
          { "$ref": "#/components/parameters/CreatedAfter" },
          { "$ref": "#/components/parameters/CreatedBefore" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "running",
                "waiting_for_input",
                "completed",
                "failed",
                "cancel_requested",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin-scoped connector runs",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/run-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/failures": {
      "get": {
        "operationId": "adminListFailures",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/AgentRef" },
          { "$ref": "#/components/parameters/CreatedAfter" },
          { "$ref": "#/components/parameters/CreatedBefore" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Failed connector runs with safe error codes",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/failure-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/admin/handoffs": {
      "get": {
        "operationId": "adminListHandoffs",
        "security": [{ "adminToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SiteId" },
          { "$ref": "#/components/parameters/AgentRef" },
          { "$ref": "#/components/parameters/CreatedAfter" },
          { "$ref": "#/components/parameters/CreatedBefore" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["requested", "awaiting_contact", "delivering", "completed", "failed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin-scoped handoffs",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/admin/handoff-list.schema.json" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionToken": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Separately signed tenant- and site-bound admin token"
      },
      "serviceCredential": { "type": "apiKey", "in": "header", "name": "X-Formation-Service-Key" }
    },
    "parameters": {
      "SiteId": {
        "name": "siteId",
        "in": "query",
        "schema": { "$ref": "../schemas/common/opaque-id.schema.json" }
      },
      "AgentRef": {
        "name": "agentRef",
        "in": "query",
        "schema": { "$ref": "../schemas/common/opaque-id.schema.json" }
      },
      "CreatedAfter": {
        "name": "createdAfter",
        "in": "query",
        "schema": { "$ref": "../schemas/common/timestamp.schema.json" }
      },
      "CreatedBefore": {
        "name": "createdBefore",
        "in": "query",
        "schema": { "$ref": "../schemas/common/timestamp.schema.json" }
      },
      "ConversationId": {
        "name": "conversationId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "../schemas/common/opaque-id.schema.json" }
      },
      "RequestId": {
        "name": "requestId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "../schemas/common/opaque-id.schema.json" }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "$ref": "../schemas/common/cursor-page-request.schema.json#/properties/cursor"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "LastEventId": {
        "name": "Last-Event-ID",
        "in": "header",
        "schema": { "$ref": "../schemas/common/opaque-id.schema.json" }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Scoped retry key. Same key and payload returns the original result; same key with a different payload returns 409 IDEMPOTENCY_CONFLICT.",
        "schema": {
          "$ref": "../schemas/common/idempotency-metadata.schema.json#/properties/idempotencyKey"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Stable error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "../schemas/common/error-envelope.schema.json" }
          }
        }
      }
    }
  }
}